Chore : 의미없는 코드들 삭제
