+++ JuiceMaker/JuiceMaker/Alerts.swift
@@ -10,9 +10,7 @@ import UIKit
 extension UIViewController {
     func successAlert(_ juice: Juice) {
         let alert = UIAlertController(title: "주문 확인", message: "\(juice.name) 나왔습니다! 맛있게 드세요!", preferredStyle: .alert)
-        //FIXME: 여기다 self.update()하면 label 업데이트가 안된다.. 왤까
-        
-        
+
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.85) {
             alert.dismiss(animated: true, completion: nil)
         }
@@ -21,15 +19,16 @@ extension UIViewController {
 
     func failAlert() {
         let failAlert = UIAlertController(title: nil , message: "재료가 모자라요 재고를 수정할까요?", preferredStyle: .alert)
-        let cancel = UIAlertAction(title: "아니오", style: .default)
-        let stockSettingAction = UIAlertAction(title: "예", style: .cancel)
+        let yesAction = UIAlertAction(title: "예", style: .cancel)
         { action in
             let main = UIStoryboard(name: "Main", bundle: nil)
             let stockView = main.instantiateViewController(withIdentifier: "StockViewController")
             self.present(stockView, animated: false, completion: nil)
         }
-        failAlert.addAction(stockSettingAction)
-        failAlert.addAction(cancel)
+        let noAction = UIAlertAction(title: "아니오", style: .default)
+        
+        failAlert.addAction(yesAction)
+        failAlert.addAction(noAction)
         present(failAlert, animated: true, completion: nil)
     }
 }
