+++ JuiceMaker/JuiceMaker/Base.lproj/Main.storyboard
@@ -8,10 +8,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--맛있는 쥬스를 만들어 드려요!-->
+        <!--Main-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController storyboardIdentifier="Main" id="BYZ-38-t0r" customClass="ViewController" customModule="JuiceMaker" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="Main" id="BYZ-38-t0r" userLabel="Main" customClass="ViewController" customModule="JuiceMaker" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="896" height="414"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -426,7 +426,7 @@
                                 <rect key="frame" x="776" y="10" width="56" height="30"/>
                                 <state key="normal" title="재고 수정"/>
                                 <connections>
-                                    <segue destination="RWe-6J-gOF" kind="show" animates="NO" id="60N-9N-EvT"/>
+                                    <segue destination="RWe-6J-gOF" kind="presentation" animates="NO" id="60N-9N-EvT"/>
                                 </connections>
                             </button>
                         </subviews>
@@ -642,7 +642,7 @@
                                     <constraint firstAttribute="height" constant="50" id="Pcu-sn-jrv"/>
                                 </constraints>
                                 <attributedString key="attributedText">
-                                    <fragment content="맛있는 쥬스를 만들어 드려요!">
+                                    <fragment content="재고 추가">
                                         <attributes>
                                             <color key="NSColor" name="textColor" catalog="System" colorSpace="catalog"/>
                                             <font key="NSFont" metaFont="system" size="24"/>
@@ -677,7 +677,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="pWd-tb-uOl" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="1307.8125" y="115.94202898550725"/>
+            <point key="canvasLocation" x="1367" y="116"/>
         </scene>
     </scenes>
     <designables>
