+++ JuiceMaker/JuiceMaker/FruitClass.swift
